+++ executorCommon/src/main/scala/dx/executor/TaskExecutor.scala
@@ -68,10 +68,10 @@ abstract class TaskExecutor(jobMeta: JobMeta,
   }
 
   /**
-    * Returns the IR type and value for each task input, including default values
-    * for any missing optional parameters.
+    * Returns the IR type and value for all task input variables, including default values for any
+    * missing optional parameters.
     */
-  protected def getInputsWithDefaults: Map[DxName, (Type, Value)]
+  protected def getInputVariables: Map[DxName, (Type, Value)]
 
   /**
     * Returns the minimal (i.e. cheapest) instance type that is
@@ -243,8 +243,8 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     }
   }
 
-  private val streamingDir = workerPaths.getDxfuseMountDir()
-  private val localizationDirs = Vector(workerPaths.getInputFilesDir(), streamingDir)
+  private val streamingDir = workerPaths.getDxfuseMountDir().asJavaPath
+  private val localizationDirs = Vector(workerPaths.getInputFilesDir().asJavaPath, streamingDir)
 
   private class InputFinalizer(
       uriToPath: Map[String, Path],
@@ -376,7 +376,7 @@ abstract class TaskExecutor(jobMeta: JobMeta,
   protected def writeCommandScript(
       localizedInputs: Map[DxName, (Type, Value)],
       localizedDependencies: Option[Map[String, (Type, Value)]]
-  ): (Map[DxName, (Type, Value)], Boolean, Option[Set[Int]])
+  ): (Boolean, Option[Set[Int]])
 
   /**
     * Evaluates the outputs of the task. Returns mapping of output parameter
@@ -537,7 +537,7 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     }
 
     def addVirtualFile(content: String, uri: String, destination: String): Unit = {
-      val localPath = workerPaths.getVirtualFilesDir(ensureExists = true).resolve(uri)
+      val localPath = workerPaths.getVirtualFilesDir(ensureExists = true).resolve(uri).asJavaPath
       val canonicalPath = FileUtils.writeFileContent(localPath, content)
       virtualFiles += (uri -> canonicalPath)
       addFile(canonicalPath, destination)
@@ -809,7 +809,7 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     //  we rely on DockerUtils to download the image (via DxFileSource, which uses the dx API
     //  to download the file).
     val pathsToLocalize = new PathsToLocalize
-    val inputs = pathsToLocalize.updateListingsAndExtractFiles(getInputsWithDefaults)
+    val inputs = pathsToLocalize.updateListingsAndExtractFiles(getInputVariables)
 
     if (checkInstanceType) {
       // calculate the required instance type
@@ -864,7 +864,8 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     // localize virtual files to /home/dnanexus/virtual
     trace(s"Virtual files = ${pathsToLocalize.virtualFiles}")
     val virtualUriToPath = pathsToLocalize.virtualFiles.map { fs =>
-      val localPath = fs.localizeToDir(workerPaths.getVirtualFilesDir(ensureExists = true))
+      val localPath =
+        fs.localizeToDir(workerPaths.getVirtualFilesDir(ensureExists = true).asJavaPath)
       fs.name -> localPath
     }.toMap
 
@@ -876,7 +877,7 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     // collisions in the manner specified by the WDL spec. All remote input files except those
     // streamed by dxfuse are placed in subfolders of the /home/dnanexus/inputs directory.
     val localizer = SafeLocalizationDisambiguator.create(
-        rootDir = workerPaths.getInputFilesDir(),
+        rootDir = workerPaths.getInputFilesDir().asJavaPath,
         separateDirsBySource = true,
         createDirs = true,
         disambiguationDirLimit = TaskExecutor.MaxDisambiguationDirs,
@@ -901,7 +902,8 @@ abstract class TaskExecutor(jobMeta: JobMeta,
         .foreach {
           case DxdaManifest(manifestJs) =>
             // write the manifest to a file
-            FileUtils.writeFileContent(workerPaths.getDxdaManifestFile(), manifestJs.prettyPrint)
+            FileUtils.writeFileContent(workerPaths.getDxdaManifestFile().asJavaPath,
+                                       manifestJs.prettyPrint)
             // run dxda via a subprocess
             jobMeta.runJobScriptFunction(TaskExecutor.DownloadDxda)
         }
@@ -915,7 +917,7 @@ abstract class TaskExecutor(jobMeta: JobMeta,
       trace(s"Files to stream: ${pathsToLocalize.filesToStream}")
       // use a different localizer for the dxfuse mount point
       val streamingLocalizer = SafeLocalizationDisambiguator.create(
-          rootDir = workerPaths.getDxfuseMountDir(),
+          rootDir = workerPaths.getDxfuseMountDir().asJavaPath,
           existingPaths = localizer.getLocalizedPaths,
           separateDirsBySource = true,
           createDirs = false,
@@ -933,7 +935,8 @@ abstract class TaskExecutor(jobMeta: JobMeta,
         })
         .foreach {
           case DxfuseManifest(manifestJs) =>
-            FileUtils.writeFileContent(workerPaths.getDxfuseManifestFile(), manifestJs.prettyPrint)
+            FileUtils.writeFileContent(workerPaths.getDxfuseManifestFile().asJavaPath,
+                                       manifestJs.prettyPrint)
             // run dxfuse via a subprocess
             jobMeta.runJobScriptFunction(TaskExecutor.DownloadDxfuse)
         }
@@ -972,7 +975,8 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     // - For streaming files, if the source container is the same as one managed by
     //   download localizer, link the streaming file into the download directory.
     val inputFinalizer = new InputFinalizer(uriToSourcePath, localizer)
-    val (finalizedInputs, localPathToUri) = inputFinalizer.finalizeInputs(inputs)
+    val (localizedInputs, localPathToUri) = inputFinalizer.finalizeInputs(inputs)
+    logFields(localizedInputs, "Localized inputs")
 
     // Finalize any static dependencies and update the task source code if necessary
     val finalizedDependencies = staticDependencies.map(inputFinalizer.finalizeStaticDependencies)
@@ -983,12 +987,10 @@ abstract class TaskExecutor(jobMeta: JobMeta,
     // Evaluate the command script and writes it to disk. Inputs are supplemented with
     // any local file paths created when evaluating the command script and are serialized
     // for use in the next phase.
-    val (localizedInputs, hasCommand, successCodes) =
-      writeCommandScript(finalizedInputs, finalizedDependencies)
-    logFields(localizedInputs, "Localized inputs")
+    val (hasCommand, successCodes) = writeCommandScript(localizedInputs, finalizedDependencies)
     if (hasCommand) {
       // run the command script
-      jobMeta.runJobScriptFunction(TaskExecutor.RunCommand, successCodes)
+      jobMeta.runJobScriptFunction(TaskExecutor.RunCommand, successCodes, truncateLogs = false)
     }
 
     // evaluate output expressions