+++ executorWdl/src/test/scala/dx/executor/wdl/WdlWorkflowExecutorTest.scala
@@ -9,7 +9,7 @@ import dx.core.io.DxWorkerPaths
 import dx.core.ir.{DxName, ParameterLinkSerializer, ParameterLinkValue, Type, TypeSerde}
 import dx.core.languages.wdl.{WdlBlock, WdlBundle, WdlDxName, WdlUtils}
 import dx.executor.{JobMeta, WorkflowAction, WorkflowExecutor}
-import dx.util.{CodecUtils, FileSourceResolver, FileUtils, Logger}
+import dx.util.{CodecUtils, FileSourceResolver, FileUtils, Logger, PosixPath}
 import dx.util.protocols.DxFileAccessProtocol
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -53,7 +53,8 @@ private case class WorkflowTestJobMeta(override val workerPaths: DxWorkerPaths,
   override val jobId: String = null
 
   override def runJobScriptFunction(name: String,
-                                    successCodes: Option[Set[Int]] = Some(Set(0))): Unit = ()
+                                    successCodes: Option[Set[Int]] = Some(Set(0)),
+                                    truncateLogs: Boolean = true): Unit = ()
 
   override val analysis: Option[DxAnalysis] = None
 
@@ -125,7 +126,8 @@ class WdlWorkflowExecutorTest extends AnyFlatSpec with Matchers {
   private def setup(): DxWorkerPaths = {
     // Create a clean temp directory for the task to use
     val jobRootDir: Path = Files.createTempDirectory("dxcompiler_applet_test")
-    val workerPaths = DxWorkerPaths(jobRootDir)
+    jobRootDir.toFile.deleteOnExit()
+    val workerPaths = DxWorkerPaths(PosixPath(jobRootDir.toString))
     workerPaths.createCleanDirs()
     workerPaths
   }
@@ -145,7 +147,7 @@ class WdlWorkflowExecutorTest extends AnyFlatSpec with Matchers {
   private def createFileResolver(workerPaths: DxWorkerPaths): FileSourceResolver = {
     val dxProtocol = DxFileAccessProtocol(dxApi)
     FileSourceResolver.create(
-        localDirectories = Vector(workerPaths.getWorkDir()),
+        localDirectories = Vector(workerPaths.getWorkDir().asJavaPath),
         userProtocols = Vector(dxProtocol),
         logger = logger
     )