+++ project/Merging.scala
@@ -43,6 +43,11 @@ object Merging {
           MergeStrategy.first
         case "maven" :: "com.google.guava" :: xs =>
           MergeStrategy.first
+        case _ :+ ("asm-license.txt" | "module-info.class" | "overview.html" |
+            "cobertura.properties") =>
+          MergeStrategy.discard
+        case _ :+ ("generated-layer.xml" | "namedservices.index" | "java.lang.runnable") =>
+          MergeStrategy.discard
         case _ =>
           val oldStrategy = (assemblyMergeStrategy in assembly).value
           oldStrategy(x)
@@ -57,7 +62,9 @@ object Merging {
           val oldStrategy = (assemblyMergeStrategy in assembly).value
           oldStrategy(x)
       }
-    case "asm-license.txt" | "module-info.class" | "overview.html" | "cobertura.properties" =>
+    case PathList(
+        "asm-license.txt" | "module-info.class" | "overview.html" | "cobertura.properties"
+        ) =>
       MergeStrategy.discard
     case PathList("mime.types") =>
       MergeStrategy.last